requirements.txt
Flask==3.1.0
Flask-Login==0.6.3
Flask-WTF==1.2.2
Flask-SQLAlchemy==3.1.1
Flask-Limiter==3.8.0
psycopg2-binary==2.9.9
ldap3==2.9.1
google-api-python-client==2.160.0
google-auth==2.38.0
msal==1.32.0
gunicorn==23.0.0
python-dotenv==1.1.0
requests==2.32.3
